Move DoActorLoad to CourseScene.cs
Everything related to displaying information about the selected level should be done in CourseScene.cs

Also got rid of some unused local variables in Program.cs
